+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './classification';
 export * from './confirmation';
 export * from './globalization';
 export * from './identification';